Split sequencer components
